feat: follow some standard typescript conventions (#15)
* feat: use env file for credentials

Rather than using a mysterious "auth/keys.js" file, the project
should use the standard .env file for handling project secrets.

* fix: load mongo uri from .env instead of hard coding

If this is changed to a different value, that change might be committed
and a contributor could accidentally expose their credentials. Instead,
you should pull it from the .env file.

* chore: remove compiled javascript

You should not commit these to version control. They aren't necessary
for contribution, and can be generated locally anyway.

* fix: ignore built javascript

Also fixes a typo in the directory name

* feat: add package.json scripts

This allows contributors to quickly build and start the codebase.
Also adds TypeScript as a dependency - in general, you don't want to
rely on a global version of TypeScript, but have a specific version
explicitly included in your project.